Move get_characters_loss
